Format was using the predefined SD object instead of instance.
@PaulStoffregen @mjs513 -

Mike was having issues where only SD.begin worked in MTP-test sketch... So after fixing to get both disks to show again on the host, and neither excepted a format, figured out you were hard codding to use the SD. object instead of using the object instance you were actaully called on.

Simple fix.  Just remove SD.

Format is now called.  And was able to format the one on Pin 10... But screwed up MTP as it took too long.  So I may need to change my handling of the Format code in MTP and always start up timer and process in background.