If we detected no SD return 0 for usedSize/totalSize
Hi @PaulStoffregen @mjs513

With the media detection code when we detect that the SD went away,  when you do a refresh of the top level Teensy object in MTP, it still returns the same values as before as the SDFat has this data cached.

This code simply detect that our cardPreviouslyPresent is set when we do begin and when we do detection so if we belive there is no SD card, simply return 0.

WIth this now if you do a refresh in MTP, the drive shows up in RED as expected.